huhu
git-svn-id: http://sdoatsidc.oma.be/svn/spoca@10 96c9e701-61e4-4e2c-872f-0a44b20625ab

+++ classes/FCMClassifier.cpp
@@ -133,7 +133,7 @@ void FCMClassifier::classification(Real precision, unsigned maxNumberIteration)
 	
 	#if DEBUG >= 2
 		stepinit(outputFileName+"iterations.txt");
-		unsigned decimals = 1 - log10(precision);
+		unsigned decimals = unsigned(1 - log10(precision));
 	#endif
 	
 	//Initialisation of precision & U
