+++ classes/HistogramPCMClassifier.cpp
@@ -179,7 +179,7 @@ void HistogramPCMClassifier::classification(Real precision, unsigned maxNumberIt
 
 	#if DEBUG >= 2
 		stepinit(outputFileName+"iterations.txt");
-		unsigned decimals = 1 - log10(precision);;
+		unsigned decimals = unsigned(1 - log10(precision));;
 	#endif
 
 	const Real maxFactor = ETA_MAXFACTOR;
